feat: 重构mesh工具类的代码组织
