+++ core/numerics/mesh/__init__.py
@@ -1,3 +1,4 @@
 from core.numerics.mesh.elements import *
 from core.numerics.mesh.meshes import *
 from core.numerics.mesh.grids import *
+from core.numerics.mesh.auxiliaries import *
